update in db insert, exceptions, split class to files
