+++ PoZiomkaDomain/Exceptions/ObjectNotFound.cs
@@ -1,0 +1,10 @@
+﻿using System.Runtime.Serialization;
+
+namespace PoZiomkaDomain.Exceptions;
+public class ObjectNotFound : Exception
+{
+	public ObjectNotFound(string exceptionMessage) : base(exceptionMessage)
+	{
+	}
+}
+
